+++ History.md
@@ -11,6 +11,7 @@
 - Antragsgrün supports now multilingual consultations, where documents are submitted in multiple languages in parallel and users can choose the language to view the page and motions in. Provisions for automatic tranlations of motions and amendments are there including one demo-plugin.
 - Motion screening can now be configured on a per-motion-type basis.
 - More uploaded PDFs can be handled without having to install additional software or libraries.
+- Proposed procedure versions can now be retrospectively deleted.
 
 ### Version 4.17.2 [not released yet]